cleaning up and refactoring
