+++ Server Backend/helpers/Debug.js
@@ -1,3 +1,4 @@
+'use strict'
 /**
  * Imports the console colors package
  */
