+++ Server Backend/helpers/Socket.js
@@ -1,2 +1,3 @@
+'use strict'
 const socketio = require('socket.io')
 module.exports = (server) => { return socketio(server) }
